update airflow image with imageupdater
